+++ app/controllers/pdl.py
@@ -66,6 +66,7 @@ RCPP_STATE_DISTRIBUTION_DATA_JSON = "rcpp_state_distribution_data.json"
 RCPP_SUBPROGRAMS_DATA_JSON = "rcpp_subprograms_data.json"
 II_HOUSE_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "house_outlay")
 HOUSE_PREDICTED_DATA_JSON = "house_outlay_max.json"
+HOUSE_PREDICTED_PRACTICE_CATEGORIES_DATA_JSON = "house_outlay_practices.json"
 TITLE_IV_DATA_PATH = os.path.join("controllers", "data", "title-iv")
 IV_SNAP_DATA_PATH = os.path.join(TITLE_IV_DATA_PATH, "programs", "snap")
 SNAP_STATE_DISTRIBUTION_DATA_JSON = "snap_state_distribution_data.json"
@@ -690,6 +691,20 @@ def titles_title_ii_programs_house_outlay_predicted_search():
     # open file
     with open(house_outlay_data, 'r') as predicted_data:
         file_data = predicted_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+# /pdl/titles/title-ii/programs/house-outlay/practice-names
+def titles_title_ii_programs_house_outlay_practice_names_search():
+    # set the file path
+    house_outlay_data = os.path.join(II_HOUSE_DATA_PATH, HOUSE_PREDICTED_PRACTICE_CATEGORIES_DATA_JSON)
+
+    # open file
+    with open(house_outlay_data, 'r') as practice_data:
+        file_data = practice_data.read()
 
         # parse file
         data_json = json.loads(file_data, object_pairs_hook=OrderedDict)